Move Equatable and Logger extensions to separate files
