+++ Tests/IntegrationTests/IntegrationTests.swift
@@ -19,7 +19,7 @@ import NIOPosix
 import Synchronization
 import Testing
 
-@testable public import MQTTNIO
+@testable import MQTTNIO
 
 #if canImport(Network)
 import NIOTransportServices
@@ -1146,49 +1146,5 @@ struct IntegrationTests {
         static func read(version: MQTTConnectionConfiguration.Version, from packet: MQTTIncomingPacket) throws -> Self {
             throw InternalError.notImplemented
         }
-    }
-}
-
-extension MQTTError: Equatable {
-    public static func == (lhs: MQTTError, rhs: MQTTError) -> Bool {
-        switch (lhs, rhs) {
-        case (.failedToConnect, .failedToConnect),
-            (.connectionClosed, .connectionClosed),
-            (.serverClosedConnection, .serverClosedConnection),
-            (.unexpectedMessage, .unexpectedMessage),
-            (.decodeError, .decodeError),
-            (.websocketUpgradeFailed, .websocketUpgradeFailed),
-            (.timeout, .timeout),
-            (.retrySend, .retrySend),
-            (.wrongTLSConfig, .wrongTLSConfig),
-            (.badResponse, .badResponse),
-            (.unrecognisedPacketType, .unrecognisedPacketType),
-            (.authWorkflowRequired, .authWorkflowRequired),
-            (.cancelledTask, .cancelledTask),
-            (.packetTooLarge, .packetTooLarge),
-            (.alreadyConnectedWithSession, .alreadyConnectedWithSession),
-            (.noSessionPresent, .noSessionPresent):
-            true
-        case (.connectionError(let lhsValue), .connectionError(let rhsValue)):
-            lhsValue == rhsValue
-        case (.reasonError(let lhsValue), .reasonError(let rhsValue)):
-            lhsValue == rhsValue
-        case (.serverDisconnection(let lhsValue), .serverDisconnection(let rhsValue)):
-            lhsValue.reason == rhsValue.reason && lhsValue.properties == rhsValue.properties
-        case (.versionMismatch(let expectedLHS, let actualLHS), .versionMismatch(let expectedRHS, let actualRHS)):
-            expectedLHS == expectedRHS && actualLHS == actualRHS
-        case (.invalidTopicFilter(let lhsValue), .invalidTopicFilter(let rhsValue)):
-            lhsValue == rhsValue
-        default:
-            false
-        }
-    }
-}
-
-extension Logger {
-    func withLogLevel(_ logLevel: Logger.Level) -> Logger {
-        var logger = self
-        logger.logLevel = logLevel
-        return logger
     }
 }
