+++ Tests/MQTTNIOTests/MQTTSubscriptionsTests.swift
@@ -92,11 +92,3 @@ struct MQTTSubscriptionsTests {
         }
     }
 }
-
-extension Logger {
-    func withLogLevel(_ logLevel: Logger.Level) -> Logger {
-        var logger = self
-        logger.logLevel = logLevel
-        return logger
-    }
-}
